rename tree slice to decisionTreeStore
